EdCareer/requirements.txt
tensorflow-cpu
bokeh==2.4.1
numpy
pandas
streamlit
joblib
pathlib
nltk